fixed encryption and decryption
